Remove doc, add LabeledVideoTensorDataset
